fix: promote step variables default values to prod
