+++ apps/api/src/app/change/e2e/promote-changes.e2e.ts
@@ -3,19 +3,15 @@ import { expect } from 'chai';
 import {
   ChangeRepository,
   EnvironmentRepository,
-  LayoutRepository,
   MessageTemplateRepository,
   NotificationGroupRepository,
   NotificationTemplateRepository,
+  EnvironmentEntity,
 } from '@novu/dal';
 import {
   ChangeEntityTypeEnum,
   ChannelCTATypeEnum,
   EmailBlockTypeEnum,
-  ITemplateVariable,
-  LayoutDescription,
-  LayoutId,
-  LayoutName,
   StepTypeEnum,
   FilterPartTypeEnum,
   TemplateVariableTypeEnum,
@@ -29,8 +25,8 @@ import {
 
 describe('Promote changes', () => {
   let session: UserSession;
+  let prodEnv: EnvironmentEntity;
   const changeRepository: ChangeRepository = new ChangeRepository();
-  const layoutRepository = new LayoutRepository();
   const notificationTemplateRepository = new NotificationTemplateRepository();
   const messageTemplateRepository: MessageTemplateRepository = new MessageTemplateRepository();
   const notificationGroupRepository: NotificationGroupRepository = new NotificationGroupRepository();
@@ -39,12 +35,11 @@ describe('Promote changes', () => {
   beforeEach(async () => {
     session = new UserSession();
     await session.initialize();
+    prodEnv = await getProductionEnvironment();
   });
 
   describe('Notification template changes', () => {
     it('should set correct notification group for notification template', async () => {
-      const prodEnv = await getProductionEnvironment();
-
       const parentGroup = await notificationGroupRepository.create({
         name: 'test',
         _environmentId: session.environment._id,
@@ -105,6 +100,93 @@ describe('Promote changes', () => {
       expect(prodVersion._notificationGroupId).to.eq(prodGroup._id);
     });
 
+    it('should promote step variables default values', async () => {
+      const testTemplate: Partial<CreateNotificationTemplateRequestDto> = {
+        name: 'test email template',
+        description: 'This is a test description',
+        notificationGroupId: session.notificationGroups[0]._id,
+        steps: [
+          {
+            template: {
+              name: 'Message Name',
+              subject: 'Test email subject',
+              content: [{ type: EmailBlockTypeEnum.TEXT, content: 'This is a {{variable}}' }],
+              type: StepTypeEnum.EMAIL,
+              variables: [
+                {
+                  name: 'variable',
+                  type: TemplateVariableTypeEnum.STRING,
+                  defaultValue: 'Test Default Value',
+                  required: false,
+                },
+              ],
+            },
+          },
+        ],
+      };
+
+      const { body } = await session.testAgent.post(`/v1/notification-templates`).send(testTemplate);
+      const notificationTemplateId = body.data._id;
+
+      await session.applyChanges({
+        enabled: false,
+      });
+
+      const prodVersion = await notificationTemplateRepository.findOne({
+        _environmentId: prodEnv._id,
+        _parentId: notificationTemplateId,
+      });
+      let prodVersionMessage = await messageTemplateRepository.findOne({
+        _environmentId: prodEnv._id,
+        _id: prodVersion?.steps[0]._templateId,
+      });
+
+      const variable = prodVersionMessage?.variables?.[0];
+      expect(variable?.name).to.eq('variable');
+      expect(variable?.type).to.eq(TemplateVariableTypeEnum.STRING);
+      expect(variable?.required).to.eq(false);
+      expect(variable?.defaultValue).to.eq('Test Default Value');
+
+      const step = body.data.steps[0];
+      const update: Partial<UpdateNotificationTemplateRequestDto> = {
+        steps: [
+          {
+            _id: step._templateId,
+            _templateId: step._templateId,
+            template: {
+              type: step?.template?.type,
+              content: step.template.content,
+              variables: [
+                {
+                  name: 'variable',
+                  type: TemplateVariableTypeEnum.STRING,
+                  defaultValue: 'New Default Value',
+                  required: true,
+                },
+              ],
+            },
+          },
+        ],
+      };
+
+      await session.testAgent.put(`/v1/notification-templates/${notificationTemplateId}`).send(update);
+
+      await session.applyChanges({
+        enabled: false,
+      });
+
+      prodVersionMessage = await messageTemplateRepository.findOne({
+        _environmentId: prodEnv._id,
+        _id: prodVersion?.steps[0]._templateId,
+      });
+
+      const updatedVariable = prodVersionMessage?.variables?.[0];
+      expect(updatedVariable?.name).to.eq('variable');
+      expect(updatedVariable?.type).to.eq(TemplateVariableTypeEnum.STRING);
+      expect(updatedVariable?.required).to.eq(true);
+      expect(updatedVariable?.defaultValue).to.eq('New Default Value');
+    });
+
     it('delete message', async () => {
       const testTemplate: Partial<CreateNotificationTemplateRequestDto> = {
         name: 'test email template',
@@ -156,8 +238,6 @@ describe('Promote changes', () => {
         enabled: false,
       });
 
-      const prodEnv = await getProductionEnvironment();
-
       const prodVersion = await notificationTemplateRepository.findOne({
         _environmentId: prodEnv._id,
         _parentId: notificationTemplateId,
@@ -188,8 +268,6 @@ describe('Promote changes', () => {
       await session.applyChanges({
         enabled: false,
       });
-
-      const prodEnv = await getProductionEnvironment();
 
       const prodVersion = await notificationTemplateRepository.findOne({
         _organizationId: session.organization._id,
@@ -271,8 +349,6 @@ describe('Promote changes', () => {
       await session.applyChanges({
         enabled: false,
       });
-
-      const prodEnv = await getProductionEnvironment();
 
       const prodVersion = await messageTemplateRepository.findOne({
         _environmentId: prodEnv._id,
@@ -381,8 +457,6 @@ describe('Promote changes', () => {
       await session.applyChanges({
         enabled: false,
       });
-
-      const prodEnv = await getProductionEnvironment();
 
       const prodVersion = await notificationTemplateRepository.find({
         _environmentId: prodEnv._id,
@@ -532,7 +606,6 @@ describe('Promote changes', () => {
 
     it('should set isBlueprint correctly', async () => {
       process.env.BLUEPRINT_CREATOR = session.organization._id;
-      const prodEnv = await getProductionEnvironment();
 
       const parentGroup = await notificationGroupRepository.create({
         name: 'test',
@@ -629,321 +702,6 @@ describe('Promote changes', () => {
       expect(changes.length).to.eq(1);
     });
   });
-
-  describe('Layout changes', () => {
-    it('should promote a new layout created to production', async () => {
-      const layoutName = 'layout-name-creation';
-      const layoutDescription = 'Amazing new layout';
-      const content = '<html><body><div>Hello {{organizationName}} {{{body}}}</div></body></html>';
-      const variables = [
-        { name: 'organizationName', type: TemplateVariableTypeEnum.STRING, defaultValue: 'Company', required: false },
-      ];
-      const isDefault = true;
-
-      const createLayoutPayload = {
-        name: layoutName,
-        description: layoutDescription,
-        content,
-        variables,
-        isDefault,
-      };
-
-      const {
-        body: {
-          data: { _id: layoutId },
-        },
-      } = await session.testAgent.post('/v1/layouts').send(createLayoutPayload);
-
-      expect(layoutId).to.be.ok;
-
-      const {
-        body: { data: devLayout },
-      } = await session.testAgent.get(`/v1/layouts/${layoutId}`);
-
-      const changes = await changeRepository.find(
-        {
-          _environmentId: session.environment._id,
-          _organizationId: session.organization._id,
-          enabled: false,
-          _entityId: layoutId,
-          type: ChangeEntityTypeEnum.DEFAULT_LAYOUT,
-        },
-        '',
-        {
-          sort: { createdAt: 1 },
-        }
-      );
-
-      expect(changes.length).to.eql(1);
-      expect(changes[0]._entityId).to.eql(layoutId);
-      expect(changes[0].type).to.eql(ChangeEntityTypeEnum.DEFAULT_LAYOUT);
-      expect(changes[0].change).to.deep.include({ op: 'add', path: ['_id'], val: layoutId });
-
-      await session.applyChanges({
-        enabled: false,
-      });
-
-      const prodEnv = await getProductionEnvironment();
-
-      const prodLayout = await layoutRepository.findOne({
-        _environmentId: prodEnv._id,
-        _parentId: layoutId,
-      });
-
-      expect(prodLayout).to.be.ok;
-      expect(prodLayout._parentId).to.eql(devLayout._id);
-      expect(prodLayout._environmentId).to.eql(prodEnv._id);
-      expect(prodLayout._organizationId).to.eql(session.organization._id);
-      expect(prodLayout._creatorId).to.eql(session.user._id);
-      expect(prodLayout.name).to.eql(layoutName);
-      expect(prodLayout.content).to.eql(content);
-      // TODO: Awful but it comes from the repository directly.
-      const { _id: _, ...prodVariables } = prodLayout.variables?.[0] as any;
-      expect(prodVariables).to.deep.include(variables[0]);
-      expect(prodLayout.contentType).to.eql(devLayout.contentType);
-      expect(prodLayout.isDefault).to.eql(isDefault);
-      expect(prodLayout.channel).to.eql(devLayout.channel);
-    });
-
-    it('should promote the updates done to a layout existing to production', async () => {
-      const layoutName = 'layout-name-update';
-      const layoutDescription = 'Amazing new layout';
-      const content = '<html><body><div>Hello {{organizationName}} {{{body}}}</div></body></html>';
-      const variables = [
-        { name: 'organizationName', type: TemplateVariableTypeEnum.STRING, defaultValue: 'Company', required: false },
-      ];
-      const isDefault = false;
-
-      const layoutId = await createLayout(layoutName, layoutDescription, content, variables, isDefault);
-
-      await session.applyChanges({
-        enabled: false,
-      });
-
-      const updatedLayoutName = 'layout-name-creation-updated';
-      const updatedDescription = 'Amazing new layout updated';
-      const updatedContent = '<html><body><div>Hello {{organizationName}}, you all {{{body}}}</div></body></html>';
-      const updatedVariables = [
-        {
-          name: 'organizationName',
-          type: TemplateVariableTypeEnum.STRING,
-          defaultValue: 'Organization',
-          required: true,
-        },
-      ];
-      const updatedIsDefault = false;
-
-      const patchLayoutPayload = {
-        name: updatedLayoutName,
-        description: updatedDescription,
-        content: updatedContent,
-        variables: updatedVariables,
-        isDefault: updatedIsDefault,
-      };
-
-      const {
-        status,
-        body: { data: patchedLayout },
-      } = await session.testAgent.patch(`/v1/layouts/${layoutId}`).send(patchLayoutPayload);
-      expect(status).to.eql(200);
-
-      const changes = await changeRepository.find(
-        {
-          _environmentId: session.environment._id,
-          _organizationId: session.organization._id,
-          enabled: false,
-          _entityId: layoutId,
-          type: ChangeEntityTypeEnum.LAYOUT,
-        },
-        '',
-        {
-          sort: { createdAt: 1 },
-        }
-      );
-
-      expect(changes.length).to.eql(1);
-      expect(changes[0]._entityId).to.eql(layoutId);
-      expect(changes[0].type).to.eql(ChangeEntityTypeEnum.LAYOUT);
-      expect(changes[0].change).to.deep.include.members([
-        {
-          op: 'update',
-          path: ['name'],
-          val: updatedLayoutName,
-          oldVal: layoutName,
-        },
-        {
-          op: 'update',
-          path: ['description'],
-          val: updatedDescription,
-          oldVal: layoutDescription,
-        },
-        {
-          op: 'update',
-          path: ['description'],
-          val: updatedDescription,
-          oldVal: layoutDescription,
-        },
-        {
-          op: 'update',
-          path: ['content'],
-          val: '<html><body><div>Hello {{organizationName}}, you all {{{body}}}</div></body></html>',
-          oldVal: '<html><body><div>Hello {{organizationName}} {{{body}}}</div></body></html>',
-        },
-        {
-          op: 'update',
-          path: ['variables', 0, 'defaultValue'],
-          val: 'Organization',
-          oldVal: 'Company',
-        },
-        {
-          op: 'update',
-          path: ['variables', 0, 'required'],
-          val: true,
-          oldVal: false,
-        },
-      ]);
-
-      await session.applyChanges({
-        enabled: false,
-      });
-
-      const prodEnv = await getProductionEnvironment();
-
-      const prodLayout = await layoutRepository.findOne({
-        _environmentId: prodEnv._id,
-        _parentId: layoutId,
-      });
-
-      expect(prodLayout).to.be.ok;
-      expect(prodLayout._parentId).to.eql(patchedLayout._id);
-      expect(prodLayout._environmentId).to.eql(prodEnv._id);
-      expect(prodLayout._organizationId).to.eql(session.organization._id);
-      expect(prodLayout._creatorId).to.eql(session.user._id);
-      expect(prodLayout.name).to.eql(updatedLayoutName);
-      expect(prodLayout.content).to.eql(updatedContent);
-      // TODO: Awful but it comes from the repository directly.
-      const { _id, ...prodVariables } = prodLayout.variables?.[0] as any;
-      expect(prodVariables).to.deep.include(updatedVariables[0]);
-      expect(prodLayout.contentType).to.eql(patchedLayout.contentType);
-      expect(prodLayout.isDefault).to.eql(updatedIsDefault);
-      expect(prodLayout.channel).to.eql(patchedLayout.channel);
-    });
-
-    it('should promote the deletion of a layout to production', async () => {
-      const layoutName = 'layout-name-deletion';
-      const layoutDescription = 'Amazing new layout';
-      const content = '<html><body><div>Hello {{organizationName}} {{{body}}}</div></body></html>';
-      const variables = [
-        { name: 'organizationName', type: TemplateVariableTypeEnum.STRING, defaultValue: 'Company', required: false },
-      ];
-      const isDefault = false;
-
-      const layoutId = await createLayout(layoutName, layoutDescription, content, variables, isDefault);
-      const {
-        body: { data: devLayout },
-      } = await session.testAgent.get(`/v1/layouts/${layoutId}`);
-
-      const changes = await changeRepository.find(
-        {
-          _environmentId: session.environment._id,
-          _organizationId: session.organization._id,
-          enabled: false,
-          _entityId: layoutId,
-          type: ChangeEntityTypeEnum.LAYOUT,
-        },
-        '',
-        {
-          sort: { createdAt: 1 },
-        }
-      );
-
-      expect(changes.length).to.eql(1);
-      expect(changes[0]._entityId).to.eql(layoutId);
-      expect(changes[0].type).to.eql(ChangeEntityTypeEnum.LAYOUT);
-      expect(changes[0].change).to.deep.include({ op: 'add', path: ['_id'], val: layoutId });
-
-      await session.applyChanges({
-        enabled: false,
-      });
-
-      const {
-        body: { data: deletedLayout },
-        status,
-      } = await session.testAgent.delete(`/v1/layouts/${layoutId}`);
-
-      expect(status).to.eql(204);
-
-      const deletionChanges = await changeRepository.find(
-        {
-          _environmentId: session.environment._id,
-          _organizationId: session.organization._id,
-          enabled: false,
-          _entityId: layoutId,
-          type: ChangeEntityTypeEnum.LAYOUT,
-        },
-        '',
-        {
-          sort: { createdAt: 1 },
-        }
-      );
-
-      expect(deletionChanges.length).to.eql(1);
-      expect(deletionChanges[0]._entityId).to.eql(layoutId);
-      expect(deletionChanges[0].type).to.eql(ChangeEntityTypeEnum.LAYOUT);
-      expect(deletionChanges[0].change).to.deep.include.members([
-        {
-          op: 'update',
-          path: ['deleted'],
-          val: true,
-          oldVal: false,
-        },
-        {
-          op: 'add',
-          path: ['isDeleted'],
-          val: true,
-        },
-      ]);
-
-      await session.applyChanges({
-        enabled: false,
-      });
-
-      const prodEnv = await getProductionEnvironment();
-
-      const prodLayout = await layoutRepository.findOne({
-        _environmentId: prodEnv._id,
-        _parentId: layoutId,
-      });
-
-      expect(prodLayout).to.not.be.ok;
-    });
-  });
-
-  async function createLayout(
-    layoutName: LayoutName,
-    layoutDescription: LayoutDescription,
-    content: string,
-    variables: ITemplateVariable[],
-    isDefault: boolean
-  ): Promise<LayoutId> {
-    const createLayoutPayload = {
-      name: layoutName,
-      description: layoutDescription,
-      content,
-      variables,
-      isDefault,
-    };
-
-    const {
-      body: {
-        data: { _id: layoutId },
-      },
-    } = await session.testAgent.post('/v1/layouts').send(createLayoutPayload);
-
-    expect(layoutId).to.be.ok;
-
-    return layoutId;
-  }
 
   async function getProductionEnvironment() {
     return await environmentRepository.findOne({
