+++ apps/web/cypress/tests/changes.spec.ts
@@ -1,5 +1,5 @@
 import { dragAndDrop } from './notification-editor';
-import { goBack } from './notification-editor/index';
+import { goBack } from './notification-editor';
 
 describe('Changes Screen', function () {
   beforeEach(function () {
@@ -40,6 +40,36 @@ describe('Changes Screen', function () {
     cy.visit('/changes');
     cy.getByTestId('change-type').contains('Template Change');
     cy.getByTestId('change-content').contains('Test Notification Title');
+  });
+
+  it('should show one change for status change and template update', function () {
+    const template = this.session.templates[0];
+
+    cy.visit('/templates/edit/' + template._id);
+    cy.waitForNetworkIdle(500);
+
+    cy.getByTestId('settings-page').click();
+    cy.waitForNetworkIdle(500);
+
+    cy.getByTestId('title').first().clear().type('Updated Title');
+    cy.getByTestId('notification-template-submit-btn').click();
+
+    cy.getByTestId('side-nav-changes-count').contains('1');
+
+    cy.getByTestId('active-toggle-switch').click({ force: true });
+    cy.getByTestId('side-nav-changes-count').contains('1');
+
+    promoteNotification();
+    switchEnvironment('Production');
+    cy.location('pathname').should('equal', '/templates');
+
+    cy.getByTestId('notifications-template').find('tbody tr').first().click({ force: true });
+
+    cy.getByTestId('settings-page').click();
+    cy.waitForNetworkIdle(500);
+    cy.getByTestId('title').first().should('have.value', 'Updated Title');
+
+    cy.getByTestId('active-toggle-switch').get('label').contains('Inactive');
   });
 
   it('should show history of changes', function () {
